Refactoring THipo 	modified:   CMakeLists.txt 	modified:   include/THipo.h 	new file:   include/THipoBank.h 	new file:   include/THipoBankParser.h 	copied:     include/THipo.h -> include/THipoItem.h 	modified:   src/THipo.cxx 	new file:   src/THipoBank.cxx 	new file:   src/THipoBankParser.cxx 	new file:   src/THipoItem.cxx
	modified:   hipo-io/src/hipo.cxx
	modified:   toohip4root/include/THipo.h
	modified:   toohip4root/include/THipoBank.h
	modified:   toohip4root/include/THipoBankParser.h
	modified:   toohip4root/include/THipoItem.h
	modified:   toohip4root/src/THipoBank.cxx
	modified:   toohip4root/src/THipoBankParser.cxx

	modified:   hipocpp/include/hipo/event.h
	modified:   hipocpp/include/hipo/reader.h
	modified:   hipocpp/src/reader.cpp
	modified:   hipocpp/src/utils.cpp
	new file:   toohip4root/include/RHipoDS.hxx
	modified:   toohip4root/include/THipoBankParser.h
	new file:   toohip4root/src/RHipoDS.cxx

	modified:   ../src/hipocpp/include/hipo/reader.h
	modified:   ../src/hipocpp/include/hipo/writer.h
	modified:   ../src/hipocpp/src/reader.cpp
	modified:   ../src/hipocpp/src/writer.cpp
	modified:   ../src/toohip4root/CMakeLists.txt

	modified:   include/hipo/record.h
	modified:   include/hipo/writer.h
	modified:   src/record.cpp
	modified:   src/writer.cpp
