+++ src/toohip4root/src/THipoItem.cxx
@@ -1,0 +1,6 @@
+#include "THipoItem.h"
+
+namespace hipo {
+
+
+}
